+++ srcs/backend/backend/accounts/apps.py
@@ -1,5 +1,5 @@
 from django.apps import AppConfig
 
 class AccountsConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'accounts'
+	default_auto_field = 'django.db.models.BigAutoField'
+	name = 'accounts'
