+++ srcs/backend/backend/accounts/forms.py
@@ -6,113 +6,113 @@ from django.core import validators
 User = get_user_model()
 
 class LoginForm(forms.Form):
-    email = forms.EmailField(max_length=254)
-    password = forms.CharField(max_length=63, widget=forms.PasswordInput)
+	email = forms.EmailField(max_length=254)
+	password = forms.CharField(max_length=63, widget=forms.PasswordInput)
 
 def validateLength(value):
-    if value.isnumeric():
-        a_string = str(value)
-        length = len(a_string)
-        if length != 6:
-            raise forms.ValidationError("Invalid length")
-    else:
-        raise forms.ValidationError("Number only")
+	if value.isnumeric():
+		a_string = str(value)
+		length = len(a_string)
+		if length != 6:
+			raise forms.ValidationError("Invalid length")
+	else:
+		raise forms.ValidationError("Number only")
 
 class OTPForm(forms.Form):
-    email = forms.EmailField(max_length=254)
-    password = forms.CharField(max_length=63, widget=forms.PasswordInput)
-    otp = forms.CharField(validators=[validateLength])
+	email = forms.EmailField(max_length=254)
+	password = forms.CharField(max_length=63, widget=forms.PasswordInput)
+	otp = forms.CharField(validators=[validateLength])
 
 
 class RegisterForm(forms.ModelForm):
-    """Form par defaut"""
+	"""Form par defaut"""
 
-    password = forms.CharField(widget=forms.PasswordInput)
-    password_2 = forms.CharField(label='Confirm Password', widget=forms.PasswordInput)
-    email = forms.EmailField(max_length=254)
+	password = forms.CharField(widget=forms.PasswordInput)
+	password_2 = forms.CharField(label='Confirm Password', widget=forms.PasswordInput)
+	email = forms.EmailField(max_length=254)
 
-    class Meta:
-        model = User
-        fields = ['name', 'email', 'login42']
-    
-    def clean_email(self):
-        """Checker si le email est deja existant dans la db"""
+	class Meta:
+		model = User
+		fields = ['name', 'email', 'login42']
+	
+	def clean_email(self):
+		"""Checker si le email est deja existant dans la db"""
 
-        email = self.cleaned_data.get('email')
-        qs = User.objects.filter(email=email)
-        if qs.exists():
-            raise forms.ValidationError("This email is already taken")
-        return email
+		email = self.cleaned_data.get('email')
+		qs = User.objects.filter(email=email)
+		if qs.exists():
+			raise forms.ValidationError("This email is already taken")
+		return email
 
-    def clean_name(self):
-        """Checker si le email est deja existant dans la db"""
+	def clean_name(self):
+		"""Checker si le email est deja existant dans la db"""
 
-        name = self.cleaned_data.get('name')
-        qs = User.objects.filter(name=name)
-        if qs.exists():
-            raise forms.ValidationError("This name is already taken")
-        return name
+		name = self.cleaned_data.get('name')
+		qs = User.objects.filter(name=name)
+		if qs.exists():
+			raise forms.ValidationError("This name is already taken")
+		return name
 
-    def clean(self):
-        """Checker si les deux mdp match"""
+	def clean(self):
+		"""Checker si les deux mdp match"""
 
-        cleaned_data = super().clean()
-        password = cleaned_data.get("password")
-        password_2 = cleaned_data.get("password_2")
-        if password is not None and password != password_2:
-            self.add_error("password_2", "Your passwords must match")
-        return cleaned_data
-    
-    def save(self, commit=True):
-        """Save the user with the hashed password"""
+		cleaned_data = super().clean()
+		password = cleaned_data.get("password")
+		password_2 = cleaned_data.get("password_2")
+		if password is not None and password != password_2:
+			self.add_error("password_2", "Your passwords must match")
+		return cleaned_data
+	
+	def save(self, commit=True):
+		"""Save the user with the hashed password"""
 
-        user = super().save(commit=False)
-        user.set_password(self.cleaned_data["password"])
-        if commit:
-            user.save()
-        return user
-    
+		user = super().save(commit=False)
+		user.set_password(self.cleaned_data["password"])
+		if commit:
+			user.save()
+		return user
+	
 class UserAdminCreationForm(forms.ModelForm):
-    """Form pour creer un nouvel user"""
-    password = forms.CharField(widget=forms.PasswordInput)
-    password_2 = forms.CharField(label='Confirm Password', widget=forms.PasswordInput)
+	"""Form pour creer un nouvel user"""
+	password = forms.CharField(widget=forms.PasswordInput)
+	password_2 = forms.CharField(label='Confirm Password', widget=forms.PasswordInput)
 
-    class Meta:
-        model = User
-        fields = ['name', 'email', 'login42']
+	class Meta:
+		model = User
+		fields = ['name', 'email', 'login42']
 
-    def clean(self):
-        """Checker si les deux mdp match"""
-        cleaned_data = super().clean()
-        password = cleaned_data.get("password")
-        password_2 = cleaned_data.get("password_2")
-        if password is not None and password != password_2:
-            self.add_error("password_2", "Your passwords must match")
-        return cleaned_data
+	def clean(self):
+		"""Checker si les deux mdp match"""
+		cleaned_data = super().clean()
+		password = cleaned_data.get("password")
+		password_2 = cleaned_data.get("password_2")
+		if password is not None and password != password_2:
+			self.add_error("password_2", "Your passwords must match")
+		return cleaned_data
 
-    def save(self, commit=True):
-        """Sauvegarde le mdp dans la db en la cachant"""
-        user = super().save(commit=False)
-        user.set_password(self.cleaned_data["password"])
-        if commit:
-            user.save()
-        return user
+	def save(self, commit=True):
+		"""Sauvegarde le mdp dans la db en la cachant"""
+		user = super().save(commit=False)
+		user.set_password(self.cleaned_data["password"])
+		if commit:
+			user.save()
+		return user
 
 
 class UserAdminChangeForm(forms.ModelForm):
-    """A form for updating users. Includes all the fields on
-    the user, but replaces the password field with admin's
-    password hash display field.
-    """
-    password = ReadOnlyPasswordHashField()
+	"""A form for updating users. Includes all the fields on
+	the user, but replaces the password field with admin's
+	password hash display field.
+	"""
+	password = ReadOnlyPasswordHashField()
 
-    class Meta:
-        model = User
-        fields = ['name', 'email', 'login42', 'admin']
+	class Meta:
+		model = User
+		fields = ['name', 'email', 'login42', 'admin']
 
-    def clean_password(self):
-        # Regardless of what the user provides, return the initial value.
-        # This is done here, rather than on the field, because the
-        # field does not have access to the initial value
-        return self.initial["password"]
+	def clean_password(self):
+		# Regardless of what the user provides, return the initial value.
+		# This is done here, rather than on the field, because the
+		# field does not have access to the initial value
+		return self.initial["password"]
