+++ srcs/backend/backend/backend/urls.py
@@ -2,35 +2,35 @@
 URL configuration for backend project.
 
 The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/4.2/topics/http/urls/
+	https://docs.djangoproject.com/en/4.2/topics/http/urls/
 Examples:
 Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
+	1. Add an import:  from my_app import views
+	2. Add a URL to urlpatterns:  path('', views.home, name='home')
 Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
+	1. Add an import:  from other_app.views import Home
+	2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
 Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
+	1. Import the include() function: from django.urls import include, path
+	2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
 from django.urls import path, include
 from rest_framework_simplejwt.views import (
-    TokenObtainPairView,
-    TokenRefreshView,
-    TokenVerifyView,
-    TokenBlacklistView
+	TokenObtainPairView,
+	TokenRefreshView,
+	TokenVerifyView,
+	TokenBlacklistView
 )
 
 urlpatterns = [
-    path('api/admin/', admin.site.urls),
-    path('api/accounts/', include('accounts.urls')),
-    path('api/accounts/', include('django.contrib.auth.urls')),
-    path('api/user/', include('Userdb.urls')),
-    path('api/match/', include('match.urls')),
-    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
-    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
-    path('api/token/verify/', TokenVerifyView.as_view(), name='token_verify'),
-    path('api/token/blacklist/', TokenBlacklistView.as_view(), name='token_blacklist'),
+	path('api/admin/', admin.site.urls),
+	path('api/accounts/', include('accounts.urls')),
+	path('api/accounts/', include('django.contrib.auth.urls')),
+	path('api/user/', include('Userdb.urls')),
+	path('api/match/', include('match.urls')),
+	path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
+	path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
+	path('api/token/verify/', TokenVerifyView.as_view(), name='token_verify'),
+	path('api/token/blacklist/', TokenBlacklistView.as_view(), name='token_blacklist'),
 ]
