+++ srcs/web/www/frontend/js/pages/error/error.js
@@ -1,0 +1,5 @@
+import { loadContent } from "../../api/fetch.js";
+
+export function renderNotFound() {
+	loadContent('content', '/frontend/js/pages/error/pageNotFound.html');
+}
