--- srcs/web/www/frontend/js/pages/home/home.html
@@ -1,1 +1,0 @@
-<p>Sainte-viagra de christie de charogne de purée d'étole de boswell de verrat de tabarouette de mosus de câline de bine de géritole de crucifix de sacristi de viarge de maudine de bout d'viarge d'enfant d'chienne de cossin de viande à chien de maudit de saint-sacrament de cimonaque de ciarge de patente à gosse de batince de cibouleau de mautadit de bâtard de saint-cimonaque d'ostie de câlisse de sacrament de baptême de colon de mautadine de sacréfice de saint-ciboire de mangeux d'marde de tabarslaque de cibole.</p>
