+++ srcs/web/www/frontend/js/router.js
@@ -1,11 +1,14 @@
-import { getInfo, getRequest, loadContent, loadContentLang, postAuth, postRequest } from "./api/fetch.js";
-import { renderHome } from "./pages/home/home.js";
-import { renderLogin } from "./pages/login/login.js";
+import { getInfo, getRequest, loadContentLang, postAuth } from "./api/fetch.js";
 import { renderHeader } from "./components/header/header.js";
 import { renderFooter } from "./components/footer/footer.js";
+import { renderTemplate } from "./pages/template/template.js";
+import { renderLogin } from "./pages/login/login.js";
 import { renderSignup } from "./pages/signUp/signup.js";
+import { renderNotFound } from "./pages/error/error.js";
 
 import { Game } from '../../Game/core/Game.js';
+
+const game = new Game();
 
 export function debounce(func, delay) {
 	let timeoutId;
@@ -27,48 +30,6 @@ export function getCookie(name) {
 	if (parts.length === 2) return parts.pop().split(';').shift();
 }
 
-export async function renderTemplate() {
-	await loadContent('content', '/frontend/js/pages/template/template.html');
-	changeLanguage(localStorage.getItem("lang"));
-
-	const tooltipTriggerList = document.querySelectorAll('[data-bs-toggle="tooltip"]')
-	const tooltipList = [...tooltipTriggerList].map(tooltipTriggerEl => new bootstrap.Tooltip(tooltipTriggerEl))
-
-	document.querySelectorAll('input[name="options"]').forEach((elem) => {
-		elem.addEventListener('change', function() {
-			const nbOfGame = document.getElementById('numberOfGame')
-			if (elem.value != 'tournament') {
-				document.getElementById('reverseMode').hidden = false;
-				nbOfGame.max = 3;
-			}
-			else if (elem.value == 'tournament') {
-				nbOfGame.max = nbOfGame.nextElementSibling.value = nbOfGame.value = 1;
-				document.getElementById('reverse').checked = false;
-				document.getElementById('reverseMode').hidden = true;
-			}
-		})
-	})
-
-	document.getElementById('gameBtn').addEventListener('click', (e) => {
-		e.preventDefault();
-		let formData = new FormData(document.getElementById('gameSet'));
-		console.log("data");
-		for(let pair of formData.entries()){
-			console.log(pair[0], pair[1]);
-		}
-
-		if (formData.get('options') == 4)
-			new bootstrap.Modal(document.getElementById('lobby1', {})).show();
-		else if (formData.get('options') == 'tournament')
-			new bootstrap.Modal(document.getElementById('lobbyTournament', {})).show();
-		else
-			new bootstrap.Modal(document.getElementById('lobby', {})).show();
-	});
-}
-
-function renderNotFound() {
-	loadContent('content', '/frontend/js/pages/error/pageNotFound.html');
-}
 const routeHandlers = {
 	'': renderTemplate,
 	'#/login': renderLogin,
@@ -90,9 +51,6 @@ window.addEventListener('DOMContentLoaded', async () => {
 	await renderFooter();
 	await renderHeader();
 	await showFriendList();
-	// loadContentLang('body', document.documentElement.lang, () => {
-	// 	attachEventListeners();
-	// });
 	localStorage.setItem("lang", lang);
 	if (localStorage.getItem("lang") != 'fr' && localStorage.getItem("lang") != 'en' && localStorage.getItem("lang") != 'ja') {
 		if (navigator.language == 'fr' || navigator.language == 'en' || navigator.language == 'ja')
@@ -101,7 +59,6 @@ window.addEventListener('DOMContentLoaded', async () => {
 			localStorage.setItem("lang", "en");
 	}
 	changeLanguage(localStorage.getItem("lang"));
-	const game = new Game();
 	game.start();
 	await isLogin()
 });
@@ -122,6 +79,7 @@ async function isLogin() {
 		console.log('allo')
 	}
 }
+
 export function hideAndSeek(boolean) {
 	document.getElementById('header').hidden = boolean;
 	document.getElementById('content').hidden = boolean;
@@ -132,7 +90,6 @@ window.addEventListener('hashchange', () => {
 	hideAndSeek(false);
 	handleRoutes();
 });
-
 
 export function changeLanguage(lang) {
 	debouncedChangeLanguage(lang);
@@ -166,7 +123,6 @@ function attachEventListeners() {
 export async function showFriendList() {
 	try {
 		const userList = await getRequest('https://127.0.0.1/api/user/user_list/', {method: 'GET', headers: {'Authorization': `Bearer ${sessionStorage.getItem('access_token')}`}});
-		// console.log(userList)
 		const nbOfUsers = userList['length'];
 		let divUser = document.getElementById('friendList');
 		if (divUser.childElementCount === nbOfUsers)
@@ -197,108 +153,3 @@ export async function showFriendList() {
 		console.error(error);
 	}
 }
-
-// export async function showFriendList() {
-// 	try {
-// 		const user = await getRequest('https://127.0.0.1/api/user/user_login/');
-// 		if (!user.name)
-// 			return ;
-// 		const id = await getRequest('https://127.0.0.1/api/user/user_details/', user['name']);
-// 		const friend = await getRequest('https://127.0.0.1/api/user/user_relation/', id['id']);
-// 		const userList = await getRequest('https://127.0.0.1/api/user/user_list/');
-// 		const nbOfUsers = userList['length'];
-// 		const nbOfFriend = friend['length'];
-
-// 		let divFriend = document.getElementById('friendList');
-// 		let divAddFriend = document.getElementById('addFriend');
-// 		while (divFriend.hasChildNodes())
-// 			divFriend.removeChild(divFriend.firstChild);
-// 		while (divAddFriend.hasChildNodes())
-// 			divAddFriend.removeChild(divAddFriend.firstChild);
-
-// 		for (let i = 0; i < nbOfUsers; i++) {
-// 			let button = document.createElement('button');
-// 			let card = document.createElement('div');
-// 			let cardBody = document.createElement('div');
-			
-// 			let img = document.createElement('img');
-// 			let imgAdd = document.createElement('img');
-// 			img.src = '/frontend/img/person-fill-dash.svg';
-// 			imgAdd.src = '/frontend/img/person-fill-add.svg';
-			
-// 			let text = '';
-// 			let friendName = '';
-// 			let isFriend = false;
-// 			if (userList[i]['id'] == id['id']) {
-// 				continue ;
-// 			}
-// 			else if (nbOfFriend) {
-// 				for (let j = 0; j < nbOfFriend; j++) {
-// 					if (userList[i]['id'] == friend[j]['User1Id'] || userList[i]['id'] == friend[j]['User2Id']) {
-// 						friendName = await getRequest('https://127.0.0.1/api/user/user_details/', userList[i]['id']);
-// 						text = document.createTextNode(friendName['name']);
-// 						isFriend = true;
-// 						break ;
-// 					}
-// 				}
-// 				if (isFriend === false) {
-// 					friendName = await getRequest('https://127.0.0.1/api/user/user_details/', userList[i]['id']);
-// 					text = document.createTextNode(friendName['name']);
-// 				}
-// 			}
-// 			else {
-// 				friendName = await getRequest('https://127.0.0.1/api/user/user_details/', userList[i]['id']);
-// 				text = document.createTextNode(friendName['name']);
-// 				isFriend = false;
-// 			}
-// 			button.classList.add('btn');
-// 			cardBody.classList.add('card-body');
-// 			cardBody.classList.add('d-flex');
-// 			cardBody.classList.add('justify-content-between');
-// 			card.classList.add('card');
-// 			card.appendChild(cardBody);
-// 			cardBody.appendChild(text);
-// 			cardBody.appendChild(button);
-// 			if (isFriend === true) {
-// 				button.appendChild(img);
-// 				divFriend.appendChild(card);
-// 			}
-// 			else if (isFriend === false) {
-// 				button.addEventListener('click', () => addFriend(user['name'], userList[i]['name']));
-// 				button.appendChild(imgAdd);
-// 				divAddFriend.appendChild(card);
-// 			}
-// 		}
-// 	}
-// 	catch (error) {
-// 		console.error(error);
-// 	}
-// }
-
-// async function addFriend(userName, friendName) {
-// 	fetch('https://127.0.0.1/api/user/add_friend/' + userName + '/' + friendName + '/');
-
-// 	let divAddFriend = document.getElementById('addFriend');
-// 	let i = 0;
-// 	while (divAddFriend.children[i].children[0].childNodes[0].nodeValue != friendName)
-// 		i++;
-// 	if (divAddFriend.children[i].children[0].childNodes[0].nodeValue == friendName)
-// 		divAddFriend.removeChild(divAddFriend.children[i]);
-
-// 	let divFriend = document.getElementById('friendList');
-// 	let button = document.createElement('button');
-// 	let card = document.createElement('div');
-// 	let cardBody = document.createElement('div');
-// 	let img = document.createElement('img');
-// 	img.src = '/frontend/img/person-fill-dash.svg';
-// 	button.classList.add('btn');
-// 	cardBody.classList.add('card-body');
-// 	cardBody.classList.add('d-flex');
-// 	cardBody.classList.add('justify-content-between');
-// 	card.classList.add('card');
-// 	card.appendChild(cardBody);
-// 	cardBody.appendChild(document.createTextNode(friendName));
-// 	cardBody.appendChild(button);
-// 	button.appendChild(img);
-// 	divFriend.appendChild(card);
-// }
